feat: implement N1QL update query builder
Added N1QLUpdateQueryBuilder to facilitate building
update queries for Couchbase, enhancing query capabilities
and improving code organization.
